Changed the WebAPI to call engine.run_jobs [skip hazardlib]
