+++ openquake/server/views.py
@@ -544,7 +544,7 @@ def calc_run(request):
 
     user = utils.get_user(request)
     try:
-        job_id, pid = submit_job(inifiles[0], user, hazard_job_id)
+        pid = submit_job(inifiles[0], user, hazard_job_id).pid
     except Exception as exc:  # no job created, for instance missing .xml file
         # get the exception message
         exc_msg = str(exc)
@@ -552,7 +552,7 @@ def calc_run(request):
         response_data = exc_msg.splitlines()
         status = 500
     else:
-        response_data = dict(job_id=job_id, status='created', pid=pid)
+        response_data = dict(status='created', pid=pid)
         status = 200
     return HttpResponse(content=json.dumps(response_data), content_type=JSON,
                         status=status)
@@ -563,16 +563,10 @@ def submit_job(job_ini, username, hazard_job_id=None):
     Create a job object from the given job.ini file in the job directory
     and run it in a new process. Returns the job ID and PID.
     """
-    job_id = logs.init('job', logging.INFO)
-    oq = engine.job_from(
-        job_ini, job_id, username, hazard_calculation_id=hazard_job_id)
-    proc = Process(target=engine.run_calc,
-                   args=(job_id, oq, '', DEFAULT_LOG_LEVEL, None),
-                   kwargs=dict(hazard_calculation_id=hazard_job_id,
-                               username=username))
+    proc = Process(target=engine.run_jobs,
+                   args=([job_ini], DEFAULT_LOG_LEVEL, None, '', username))
     proc.start()
-    logs.dbcmd('update_job', job_id, {'pid': proc.pid})
-    return job_id, proc.pid
+    return proc
 
 
 @require_http_methods(['GET'])
